requirements.txt
annotated-types==0.7.0
anyio==4.9.0
audioread==3.0.1
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cssselect2==0.8.0
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
filelock==3.18.0
fonttools==4.58.1
fsspec==2025.5.1
groq==0.8.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.44.0
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.5
noisereduce==3.0.0
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydyf==0.11.0
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
regex==2024.11.6
reportlab==4.0.7
requests==2.32.3
scikit-learn==1.6.1
scipy==1.11.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
torch==2.3.1
tqdm==4.66.1
triton==2.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
wheel==0.45.1
zopfli==0.2.3.post1
